Separate AutocompleteMultiSelect from form component
